Transform PPMS errors into exceptions.
