+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/PpmsResponse.java
@@ -1,0 +1,27 @@
+package gov.va.api.health.providerdirectory.service;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+public interface PpmsResponse {
+
+  Error error();
+
+  String odataContext();
+
+  @Data
+  @Builder
+  @NoArgsConstructor(access = AccessLevel.PRIVATE)
+  @AllArgsConstructor(access = AccessLevel.PRIVATE)
+  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+  public static final class Error {
+
+    private String code;
+
+    private String message;
+  }
+}
