+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/ProviderSpecialtiesResponse.java
@@ -14,11 +14,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public final class ProviderSpecialtiesResponse {
+public final class ProviderSpecialtiesResponse implements PpmsResponse {
   @JsonProperty("@odata.context")
   private String odataContext;
 
   private List<Value> value;
+
+  private Error error;
 
   /** Lazy getter. */
   public List<Value> getValue() {
